Fixing USB handling

+++ bin/dev-environments-mcus-udev.sh
@@ -10,7 +10,7 @@ if [ ! -z "$(docker ps -qf name=dev-environments-mcus)" ]; then
     docker exec -u 0 dev-environments-mcus chmod -R 777 $DEVNAME
     echo "Adding $DEVNAME to docker" >> /tmp/docker_tty.log
 
-  else
+  elif [ "$ACTION" == "remove" ]; then
 
     docker exec -u 0 dev-environments-mcus rm $DEVNAME
     echo "Removing $DEVNAME from docker" >> /tmp/docker_tty.log
